requirements.txt
Django==5.2.4
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
django-environ==0.12.0
django-cors-headers==4.7.0
djoser==2.3.3
django-filter==25.1
django-silk==5.4.0
